// web/nestspend-web/src/app/core/api/models/import-transaction-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ImportError } from '../models/import-error';

/**
 * Response for bulk import operation showing results
 */
export interface ImportTransactionResponse {

  /**
   * Number of transactions successfully created
   */
  createdCount?: number;

  /**
   * Number of transactions that failed to import
   */
  errorCount?: number;

  /**
   * List of error messages for failed imports
   */
  errors?: Array<ImportError>;

  /**
   * Number of transactions skipped (duplicates)
   */
  skippedCount?: number;
}